CV_MIND-main/CV_MIND-main/requirements.txt
Authlib==1.3.1
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
python-dotenv==1.2.1
pytz==2024.1
requests==2.32.5
Werkzeug==3.1.3
pymongo==4.16.0
